Refactor: Move message classes to new package
