+++ core/src/main/java/bisq/core/proto/network/CoreNetworkProtoResolver.java
@@ -46,8 +46,6 @@ import bisq.core.support.dispute.messages.OpenNewDisputeMessage;
 import bisq.core.support.dispute.messages.PeerOpenedDisputeMessage;
 import bisq.core.support.dispute.refund.refundagent.RefundAgent;
 import bisq.core.support.messages.ChatMessage;
-import bisq.core.trade.messages.CancelTradeRequestAcceptedMessage;
-import bisq.core.trade.messages.CancelTradeRequestRejectedMessage;
 import bisq.core.trade.messages.CounterCurrencyTransferStartedMessage;
 import bisq.core.trade.messages.DelayedPayoutTxSignatureRequest;
 import bisq.core.trade.messages.DelayedPayoutTxSignatureResponse;
@@ -60,8 +58,10 @@ import bisq.core.trade.messages.MediatedPayoutTxSignatureMessage;
 import bisq.core.trade.messages.PayoutTxPublishedMessage;
 import bisq.core.trade.messages.PeerPublishedDelayedPayoutTxMessage;
 import bisq.core.trade.messages.RefreshTradeStateRequest;
-import bisq.core.trade.messages.RequestCancelTradeMessage;
 import bisq.core.trade.messages.TraderSignedWitnessMessage;
+import bisq.core.trade.messages.cancel.CancelTradeRequestAcceptedMessage;
+import bisq.core.trade.messages.cancel.CancelTradeRequestRejectedMessage;
+import bisq.core.trade.messages.cancel.RequestCancelTradeMessage;
 import bisq.core.trade.statistics.TradeStatistics;
 
 import bisq.network.p2p.AckMessage;
